StreamingContextManager now a stand alone class
