+++ spinn_front_end_common/utility_models/__init__.py
@@ -29,6 +29,7 @@ from .multi_cast_command import MultiCastCommand
 from .reverse_ip_tag_multi_cast_source import ReverseIpTagMultiCastSource
 from .reverse_ip_tag_multicast_source_machine_vertex import (
     ReverseIPTagMulticastSourceMachineVertex)
+from .streaming_context_manager import StreamingContextManager
 
 __all__ = ["CommandSender", "CommandSenderMachineVertex",
            "ChipPowerMonitor", "ChipPowerMonitorMachineVertex",
@@ -36,4 +37,5 @@ __all__ = ["CommandSender", "CommandSenderMachineVertex",
            "ExtraMonitorSupport", "ExtraMonitorSupportMachineVertex",
            "LivePacketGather", "LivePacketGatherMachineVertex",
            "MultiCastCommand", "ReverseIpTagMultiCastSource",
-           "ReverseIPTagMulticastSourceMachineVertex"]
+           "ReverseIPTagMulticastSourceMachineVertex",
+           "StreamingContextManager"]
